Generate job groups on the character selection screen from yaml files
This commit adds the following new, automatically generated file:
 - job-groups.h
generated by job-gen.py, using almost identical code to that for the
species groups generated by species-gen.py.

Given how much of the code in job-gen.py is identical to that in
species-gen.py, a subsequent commit will aim to move functions that
are very similar/the same into a central utility .py file.