Move some common functions out of job-gen and species-gen
This commit moves common validation and other helper functions that
would be useful for any yaml-to-header generation file out of
job-gen.py and species-gen.py, and into a new file gen_util.py,
and lets the two scripts import gen_util.py.

Note the name of the utility module is gen_util.py with an underscore,
because importing file names containing the hyphen character is
impractical in Python.